requirements.txt
django==4.0.7
django-debug-toolbar==3.2.1
requests==2.28.0
python-dotenv==0.20.0
django-phonenumber-field[phonenumbers]==6.1.0
djangorestframework==3.13.1
celery==5.2.7
redis==4.3.4
drf-yasg==1.21.3
django-celery-beat==2.3.0
django-celery-results==2.4.0
python-decouple==3.6